integrate fixtures into conftest
