+++ tests/data/tavern/api_version.yaml
@@ -7,11 +7,11 @@ strict:
 stages:
   - name: version (json output)
     request:
-      url: http://{daemon_host}:{daemon_port}/sabnzbd/api
+      url: http://{SAB_HOST}:{SAB_PORT}/sabnzbd/api
       method: GET
       params:
         mode: version
-        apikey: "{daemon_apikey}"
+        apikey: "{SAB_APIKEY}"
         output: json
     response:
       status_code: 200
@@ -19,15 +19,15 @@ stages:
         content-type: !re_match "application/json"
         content-type: !re_search "charset=(UTF|utf)-8"
       json:
-        version: "{daemon_version}"
+        version: "{SAB_VERSION}"
 
   - name: version (text output)
     request:
-      url: http://{daemon_host}:{daemon_port}/sabnzbd/api
+      url: http://{SAB_HOST}:{SAB_PORT}/sabnzbd/api
       method: GET
       params:
         mode: version
-        apikey: "{daemon_apikey}"
+        apikey: "{SAB_APIKEY}"
         output: text
     response:
       status_code: 200
@@ -37,15 +37,15 @@ stages:
       verify_response_with:
         function: tavern.testutils.helpers:validate_regex
         extra_kwargs:
-          expression: "{daemon_version}"
+          expression: "{SAB_VERSION}"
 
   - name: version (xml output)
     request:
-      url: http://{daemon_host}:{daemon_port}/sabnzbd/api
+      url: http://{SAB_HOST}:{SAB_PORT}/sabnzbd/api
       method: GET
       params:
         mode: version
-        apikey: "{daemon_apikey}"
+        apikey: "{SAB_APIKEY}"
         output: xml
     response:
       status_code: 200
@@ -55,4 +55,4 @@ stages:
       verify_response_with:
         function: tavern.testutils.helpers:validate_regex
         extra_kwargs:
-          expression: '<\?xml version="1.0" encoding="UTF-8" \?>\r?\n?<version>{daemon_version}</version>'
+          expression: '<\?xml version="1.0" encoding="UTF-8" \?>\r?\n?<version>{SAB_VERSION}</version>'
